+++ test/testdata/top_level_reimports/_internal.py
@@ -1,0 +1,11 @@
+class Foo:
+    class FooSub:
+        pass
+
+
+class Bar(Foo):
+    pass
+
+
+def baz(f: Foo) -> Foo.FooSub:
+    pass