+++ pdoc/__init__.py
@@ -377,7 +377,7 @@ You can find an example in [`examples/library-usage`](https://github.com/mitmpro
 '''
 from __future__ import annotations
 
-__version__ = "10.0.1"  # this is read from setup.py
+__version__ = "10.0.2"  # this is read from setup.py
 
 import traceback
 import warnings