+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased: pdoc next
 
+ - Improve linking of classes that are re-exported in a common top-level namespace.
  - Make it more clear that Markdown ist the default docformat. ([@Dliwk](https://github.com/Dliwk))
 
 # 2022-02-14: pdoc 10.0.1